requirements.txt
requests
lxml
boto3
psycopg2-binary
pyspark
python-dotenv
kafka-python
great_expectations
prometheus-client
